refactor: board 패키지 구조를 나머지 도메인과 통일
homework/attachment/reply 와 동일하게 controller/service/repository/domain/dto
5개 패키지로 정리한다.

  application    -> service, repository
  infrastructure -> repository, domain
  presentation   -> controller, dto

## 통합하지 않고 구조만 맞춘 이유

다른 도메인과 달리 board 의 도메인 모델은 살아 있다. 13개 행위 메서드 중
9개가 실제로 호출된다(changeBoardType, increaseFavorite, checkAndSetRecommended,
validateNoticePermission 등). homework 는 7개 중 5개가 죽어 있어서 통합이
정당했지만 board 는 그렇지 않다.

Board 는 불변 모델이라 JPA 엔티티와 합치면 9개 메서드가 전부 가변으로 바뀌고
호출부 약 20곳이 영향받는다. 얻는 것은 파일 수 감소뿐이고 잃는 것은 실제로
쓰이는 도메인 로직의 불변성이라 통합하지 않았다.

## 서비스 인터페이스는 제거

BoardService / BoardCategoryService 인터페이스는 구현체가 하나뿐이라 제거하고
Impl 을 승격했다. default 메서드 2개씩은 구현체로 옮겼다.

## 이슈 #15 의 사실 정정

이슈에 BoardSearchRepository 를 "Spring Data custom repository fragment 라
제거하면 안 된다"고 적었는데 틀렸다. BoardJpaRepository 는 이 인터페이스를
extends 하지 않는다. @Repository 빈으로 BoardRepositoryImpl 에 직접 주입되는
일반 인터페이스다. 다만 board 는 Port/Adapter 를 유지하기로 했으므로 이번
변경 범위 밖이다.

테스트 166개 전부 통과.
