+++ src/main/java/com/edu/edumeet/board/dto/BoardCategoryDTO.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.board.presentation.dto;
+package com.edu.edumeet.board.dto;
 
 import lombok.*;
