+++ src/main/java/com/edu/edumeet/board/dto/BoardListAllDTO.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.board.presentation.dto;
+package com.edu.edumeet.board.dto;
 
 import com.edu.edumeet.attachment.dto.AttachmentDTO;
 import com.edu.edumeet.board.domain.BoardType;
