+++ src/main/java/com/edu/edumeet/board/dto/PageResponseDTO.java
@@ -1,4 +1,4 @@
-package com.edu.edumeet.board.presentation.dto;
+package com.edu.edumeet.board.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
