+++ src/main/java/com/edu/edumeet/board/repository/BoardSearchRepository.java
@@ -1,8 +1,8 @@
-package com.edu.edumeet.board.application;
+package com.edu.edumeet.board.repository;
 
 import com.edu.edumeet.board.domain.Board;
-import com.edu.edumeet.board.presentation.dto.BoardListAllDTO;
-import com.edu.edumeet.board.presentation.dto.BoardListReplyCountDTO;
+import com.edu.edumeet.board.dto.BoardListAllDTO;
+import com.edu.edumeet.board.dto.BoardListReplyCountDTO;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
