+++ src/test/java/com/edu/edumeet/integration/board/infrastructure/BoardInfrastructureExecutionTests.java
@@ -1,12 +1,13 @@
 package com.edu.edumeet.integration.board.infrastructure;
 
-import com.edu.edumeet.board.application.BoardSearchRepository;
+import com.edu.edumeet.board.repository.BoardSearchRepository;
 import com.edu.edumeet.board.domain.Board;
 import com.edu.edumeet.board.domain.BoardCategory;
 import com.edu.edumeet.board.domain.BoardType;
-import com.edu.edumeet.board.infrastructure.*;
-import com.edu.edumeet.board.presentation.dto.BoardListAllDTO;
-import com.edu.edumeet.board.presentation.dto.BoardListReplyCountDTO;
+import com.edu.edumeet.board.domain.*;
+import com.edu.edumeet.board.repository.*;
+import com.edu.edumeet.board.dto.BoardListAllDTO;
+import com.edu.edumeet.board.dto.BoardListReplyCountDTO;
 import com.edu.edumeet.reply.repository.ReplyRepository;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.BeforeAll;
