+++ src/test/java/com/edu/edumeet/integration/board/presentation/BoardPresentationExecutionTests.java
@@ -2,14 +2,14 @@ package com.edu.edumeet.integration.board.presentation;
 
 import com.edu.edumeet.board.domain.Board;
 import com.edu.edumeet.board.domain.BoardCategory;
-import com.edu.edumeet.board.infrastructure.BoardCategoryJpaEntity;
-import com.edu.edumeet.board.infrastructure.BoardCategoryJpaRepository;
-import com.edu.edumeet.board.infrastructure.BoardJpaEntity;
-import com.edu.edumeet.board.infrastructure.BoardJpaRepository;
-import com.edu.edumeet.board.presentation.BoardService;
-import com.edu.edumeet.board.presentation.dto.BoardDTO;
-import com.edu.edumeet.board.presentation.dto.PageRequestDTO;
-import com.edu.edumeet.board.presentation.dto.PageResponseDTO;
+import com.edu.edumeet.board.domain.BoardCategoryJpaEntity;
+import com.edu.edumeet.board.repository.BoardCategoryJpaRepository;
+import com.edu.edumeet.board.domain.BoardJpaEntity;
+import com.edu.edumeet.board.repository.BoardJpaRepository;
+import com.edu.edumeet.board.service.BoardService;
+import com.edu.edumeet.board.dto.BoardDTO;
+import com.edu.edumeet.board.dto.PageRequestDTO;
+import com.edu.edumeet.board.dto.PageResponseDTO;
 import lombok.extern.log4j.Log4j2;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
