+++ src/test/java/com/edu/edumeet/integration/reply/presentation/ReplyEdgeTests.java
@@ -2,8 +2,8 @@ package com.edu.edumeet.reply.controller;
 
 import com.edu.edumeet.reply.service.ReplyService;
 import com.edu.edumeet.board.domain.Board;
-import com.edu.edumeet.board.infrastructure.BoardJpaEntity;
-import com.edu.edumeet.board.infrastructure.BoardJpaRepository;
+import com.edu.edumeet.board.domain.BoardJpaEntity;
+import com.edu.edumeet.board.repository.BoardJpaRepository;
 import com.edu.edumeet.reply.repository.ReplyRepository;
 import com.edu.edumeet.reply.dto.ReplyDTO;
 import lombok.extern.log4j.Log4j2;
